requirements.txt
flask
gunicorn
cryptography
flask-socketio
eventlet==0.30.2
Flask-Failsafe
flask_cors
mysql-connector-python
firebase-admin
requests
